Explosion resistance is now predicted! (#30654)
* First commit

* Added Network and access
