+++ Content.Server/Explosion/EntitySystems/ExplosionSystem.CVars.cs
@@ -1,8 +1,8 @@
 using Content.Shared.CCVar;
-
+using Content.Shared.Explosion.EntitySystems;
 namespace Content.Server.Explosion.EntitySystems;
 
-public sealed partial class ExplosionSystem : EntitySystem
+public sealed partial class ExplosionSystem : SharedExplosionSystem
 {
     public int MaxIterations { get; private set; }
     public int MaxArea { get; private set; }
