+++ Content.Shared/Explosion/Components/ExplosionResistanceComponent.cs
@@ -1,8 +1,8 @@
-using Content.Server.Explosion.EntitySystems;
-using Content.Shared.Explosion;
+using Content.Shared.Explosion.EntitySystems;
 using Robust.Shared.Serialization.TypeSerializers.Implementations.Custom.Prototype.Dictionary;
+using Robust.Shared.GameStates;
 
-namespace Content.Server.Explosion.Components;
+namespace Content.Shared.Explosion.Components;
 
 /// <summary>
 /// Component that provides entities with explosion resistance.
@@ -13,8 +13,8 @@ namespace Content.Server.Explosion.Components;
 ///     This is desirable over just using damage modifier sets, given that equipment like bomb-suits need to
 ///     significantly reduce the damage, but shouldn't be silly overpowered in regular combat.
 /// </remarks>
-[RegisterComponent]
-[Access(typeof(ExplosionSystem))]
+[NetworkedComponent, RegisterComponent]
+[Access(typeof(SharedExplosionSystem))]
 public sealed partial class ExplosionResistanceComponent : Component
 {
     /// <summary>
